Adds tests for ComedyCalculator, TragedyCalculator and HistoryCalculator to validade calculations
