+++ TheatricalPlayersRefactoringKata.Tests/StatmentPrinterTests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
