+++ TheatricalPlayersRefactoringKata.Tests/StatmentPrinterTests/StatementPrinterTests.TestTextStatementExample.approved.txt
