+++ TheatricalPlayersRefactoringKata.Tests/StatmentPrinterTests/StatementPrinterTests.TestXmlStatementExample.approved.txt
